+++ packages/swapkit/core/src/client/index.ts
@@ -393,7 +393,7 @@ export class SwapKitCore<T = ''> {
     assetValue: AssetValue;
     percent: number;
     from: 'sym' | 'rune' | 'asset';
-    to: 'sym' | 'rune';
+    to: 'sym' | 'rune' | 'asset';
   }) => {
     const targetAsset =
       to === 'rune'
@@ -581,7 +581,7 @@ export class SwapKitCore<T = ''> {
   connectTrezor = async (_chains: Chain, _derivationPath: number[]): Promise<void> => {
     throw new SwapKitError('core_wallet_trezor_not_installed');
   };
-  connectKeplr = async (): Promise<void> => {
+  connectKeplr = async (_chain: Chain): Promise<void> => {
     throw new SwapKitError('core_wallet_keplr_not_installed');
   };
   connectOkx = async (_chains: Chain[]): Promise<void> => {