+++ packages/wallets/keplr/CHANGELOG.md
@@ -9,6 +9,14 @@
   - @coinmasters/helpers@1.0.0
   - @coinmasters/types@1.0.0
 
+## 1.0.0-rc.2
+
+### Patch Changes
+
+- Updated dependencies [[`46830d7`](https://github.com/thorswap/SwapKit/commit/46830d7fe2f164f25466afd5d7c768022e8443bd)]:
+  - @swapkit/helpers@1.0.0-rc.1
+  - @swapkit/toolbox-cosmos@1.0.0-rc.2
+
 ## 1.0.0-rc.1
 
 ### Patch Changes