+++ packages/wallets/trezor/CHANGELOG.md
@@ -12,6 +12,22 @@
   - @coinmasters/types@1.0.0
   - @coinmasters/toolbox-utxo@1.0.0
 
+## 1.0.0-rc.3
+
+### Patch Changes
+
+- Updated dependencies [[`9dc813b`](https://github.com/thorswap/SwapKit/commit/9dc813bfad78bd8841b02fade2f96d110a1219a7)]:
+  - @swapkit/toolbox-utxo@1.0.0-rc.2
+
+## 1.0.0-rc.2
+
+### Patch Changes
+
+- Updated dependencies [[`46830d7`](https://github.com/thorswap/SwapKit/commit/46830d7fe2f164f25466afd5d7c768022e8443bd)]:
+  - @swapkit/helpers@1.0.0-rc.1
+  - @swapkit/toolbox-evm@1.0.0-rc.2
+  - @swapkit/toolbox-utxo@1.0.0-rc.1
+
 ## 1.0.0-rc.1
 
 ### Patch Changes