+++ packages/wallets/wc/CHANGELOG.md
@@ -9,6 +9,14 @@
   - @coinmasters/toolbox-cosmos@1.0.0
   - @coinmasters/types@1.0.0
 
+## 1.0.0-rc.3
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @swapkit/toolbox-cosmos@1.0.0-rc.2
+  - @swapkit/toolbox-evm@1.0.0-rc.2
+
 ## 1.0.0-rc.2
 
 ### Patch Changes